#75 [feat] api 단일화
